requirements.txt
streamlit
pandas
numpy
plotly.express
BeautifulSoup
requests
html5lib


